- Updated setup a bit (only one templating engine now) - added adodb to the repository (to make things easier and make sure the correct version is available, it will be delivered together with Flyspray by default) - database debug trace not shown by default - my searches menu fix for IE

+++ includes/class.database.php
@@ -153,20 +153,18 @@ class Database
         $ADODB_FETCH_MODE = ADODB_FETCH_ASSOC;
 
         if($this->dblink->hasTransactions === true) {
-
             $this->dblink->StartTrans();
         }
 
         if (($numrows >= 0 ) or ($offset >= 0 )) {
             $result =  $this->dblink->SelectLimit($sql, $numrows, $offset, $inputarr);
         } else {
-
            $result =  $this->dblink->Execute($sql, $inputarr);
         }
 
         if (!$result) {
 
-            if (function_exists("debug_backtrace")) {
+            if (function_exists("debug_backtrace") && defined('DEBUG_SQL')) {
                 echo "<pre style='text-align: left;'>";
                 var_dump(debug_backtrace());
                 echo "</pre>";
